+++ mkart/models/driver.py
@@ -3,7 +3,6 @@ class Driver:
     def __init__(self, id, name):
         self._id = id
         self._name = name
-        self._laps = []
 
     @property
     def name(self):
@@ -12,11 +11,3 @@ class Driver:
     @property
     def id(self):
         return self._id
-
-    @property
-    def laps(self):
-        return self._laps
-
-    @laps.setter
-    def laps(self, lap):
-        self._laps.append(lap)